GBS-6: Bumped GWT version to 2.9.0
GBS-6: Bumped GWT version to 2.9.0. I'm doing something bad here: to successfully install it locally, I'm skipping tests:

mvn clean install -DskipTests

This is not production-grade action, but gwt-bootstrap is not being deployed as a standalone library yet, and this version bump is about to be tested as integrated into RecipeLot.